Move fee model function to separate file
